requirements.txt
streamlit
pandas
matplotlib
plotly
google-generativeai
fpdf2
numpy
Pillow
python-dotenv
